+++ server.h
@@ -17,6 +17,7 @@ void receiveMsg(Mailbox<slots>* mailbox);
  * @param idx the slot's index in the underlying CircularBuffer
  * @param msg the response in the format of a struct Message
  */
-void respond(Mailbox<slots>* mailbox, int idx, Message msg);
+//void respond(Mailbox<slots>* mailbox, int idx, Message msg);
+void respond(Mailbox<slots>* mailbox, size_t idx, Message msg);
 //void respond(Mailbox* mailbox, int idx, Message msg);
